Module auth_rs::google

Validate a Google JWT received when using a Google Login

Source: Google Sign-In for Websites

Structs

CacheControl
Contains the cache control information from the key response

GoogleAuth

Jwk
A JSON Web Key, returned from Google and used to validate the JWT

MemoryCertStore
A simple in-memory cert store

Profile

Enums

Cacheability

GoogleError
All errors that may occur from using this library

Traits

CertStore